Switch to CRT string functions

+++ TextFileReader.h
@@ -4,14 +4,14 @@
 class CTextFileReader
 {
 public:
-	static const int BUF_SIZE = 512;
+	static const size_t BUF_SIZE = 512;
 	CTextFileReader();
 	~CTextFileReader();
 	bool Open(LPCTSTR path, DWORD shareMode, DWORD flagsAndAttributes);
 	void Close();
 	bool ResetPointer();
-	int ReadLine(char *text, int textMax);
-	int ReadLastLine(char *text, int textMax);
+	size_t ReadLine(char *text, size_t textMax);
+	size_t ReadLastLine(char *text, size_t textMax);
 	int Seek(int scale);
 	bool IsOpen() const { return hFile_ != INVALID_HANDLE_VALUE; }
 private:
